Implement IntoContext derive macro
