+++ core/lib/node_framework_derive/tests/ui/correct/01_from_context.rs
@@ -1,3 +1,5 @@
+#![allow(dead_code)]
+
 use zksync_node_framework_derive::FromContext;
 
 mod service {
