+++ core/lib/node_framework_derive/tests/ui/incorrect/04_into_context_default_resource.stderr
@@ -1,0 +1,5 @@
+error: `default` attribute is not allowed in `IntoContext` macro
+  --> tests/ui/incorrect/04_into_context_default_resource.rs:33:5
+   |
+33 |     _field: u8,
+   |     ^^^^^^
